+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Relax `childprocess` dependency to allow version 1.x
 * Add `CodeSpellCheck` pre-commit hook
 * Fix deadlock which was more likely to occur when setting `parallelize` on a hook to `false`
+* Fix `Mdl` hook to use JSON output and not fail on unexpected output
 
 ## 0.48.1